+++ versions/1.20.6/src/main/java/com/ayakacraft/carpetayakaaddition/mixin/rules/frostWalkerNoFreezing/FrostWalkerEnchantmentMixin.java
@@ -30,7 +30,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-@Restriction(require = @Condition(value = ModUtils.MC_ID, versionPredicates = "<1.21.1"))
+@Restriction(require = @Condition(value = ModUtils.MC_ID, versionPredicates = "<1.21"))
 @Mixin(FrostWalkerEnchantment.class)
 public class FrostWalkerEnchantmentMixin {