+++ changelog.md
@@ -1,3 +1,8 @@
 ### General
 
-- Improved command feedback for `/endermanBlockList` and fixed some i18n problems
+- Added carpet rule `teleportExpFix` transplanted from carpet-fixes
+- Support for Minecraft 26.1-snapshot-4
+
+### Bugfix
+
+- `frostWalkerNoFreezing` not working for 1.21